feat: add day test for daily workflow
